Kubernetes feature

+++ common/src/main/java/com/bank/dto/cash/UpdateBalanceRq.java
@@ -3,8 +3,10 @@ package com.bank.dto.cash;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 @AllArgsConstructor
 public class UpdateBalanceRq {
-    private Long balance;
+    private BigDecimal balance;
 }
